pipelines/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
crashtest==0.4.1
cryptography==44.0.1
distlib==0.3.9
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
kafka-python==2.0.3
keyring==25.6.0
more-itertools==10.6.0
msgpack==1.1.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pbs-installer==2025.2.12
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.1.0
poetry-core==2.1.0
py4j==0.10.9.7
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.1
RapidFuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tomlkit==0.13.2
trove-classifiers==2025.1.15.22
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.2
zipp==3.21.0
zstandard==0.23.0
